pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stocksage"
version = "0.2.0"
description = "Multi-Agent Stock Analysis System"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "openai>=1.0.0",
    "akshare>=1.10.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "yfinance>=0.2.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
stocksage = "stocksage.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
